requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Postgres
psycopg2

# Flask Framework
Flask==0.12

# Flask Packages
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-User==0.6.12
Flask-WTF==0.13.1

# Automated tests
pytest==3.0.5
pytest-cov==2.4.0

# Development tools
# Fabric3==1.13.1.post1
# tox==2.7.0

